Fix incorrect transform in apply_aspect.
x_trf goes from rawdata-space to scaled-space, so it's what should get
applied to datalims, not x_trf.inverted().  So
```
        x0, x1 = map(x_trf.inverted().transform, dL.intervalx)
        y0, y1 = map(y_trf.inverted().transform, dL.intervaly)
```
from 87c742b should have been
```
        x0, x1 = map(x_trf.transform, dL.intervalx)
        y0, y1 = map(y_trf.transform, dL.intervaly)
```
However, fixing that triggered a failure for
test_aspect_nonlinear_adjustable_datalim
which had been added in that commit, and fixing that unraveled more
issues.

The basic question is, when aspect is set and adjustable="datalim",
should we change the x limits or the y limits to get the correct aspect?
The old code used some complex conditions, which I actually haven't
managed to fully understand, to either expand or shrink one of the
axises.  Instead, just choose to always expand (rather than shrink) one
of the axises, which will avoid sending artists out-of-bounds.  (The
sole exception is in care of shared axes, which we do not touch as
explained in the comment.)

This patch caused a change in the autolimiting of
test_axes.py::test_pie_frame_grid which was buggy anyways, I forced the
old behavior by setting x/ylims manually (after checking that the
default is to expand the limits).