Make figure parameter optional when constructing canvases.
Auto-instantiating a Figure seems reasonable enough, and helps with GUI
builders.  (The auto-instantiated figure must be a plain Figure(), not
a a pyplot-registered figure, both because most likely one should not
use pyplot when building GUIs, and also more prosaically because that
would run into a circular dependency (pyplot needs a canvas already
instantiated (actually a manager that holds a canvas) to store into its
global registry).