Simplify axisartist line clipping.
axisartist.clip_path is currently used to clip the paths for curvilinear
grid lines and compute the angles at which the ticks are drawn.

We don't actually need to clip the lines (because we can just set a
clipbox on the gridline instead); actually we just need to find the
crossings positions and angles, which is much simpler, and now
encapsulated in the `_find_line_box_crossings` function.  Therefore,
deprecate the whole clip_path module.

The baseline images change slightly, mostly because we don't clip the
path ourselves anymore, but let the renderer do that; also because
floating point slop changes for ticks exactly at the edges.