Set the GTK background color to white.
This is consistent with the other backends (although this seems to be
broken on wx too), which allows semi-transparent figures to be rendered
against a consistent background (this can be checked e.g. with
`figure(facecolor=(0, 0, 0, 0))` or `figure(facecolor=(1, 1, 1, .5))`),
which would previously be rendered differently on gtk3 when using a dark
theme.

While override_background_color is marked as deprecated on gtk3, the
alternative seems to involve a bunch of CSS, and it seems unlikely to
ever be removed now that gtk development has switched to gtk4...